requirements.txt
sentence_transformers
pypdf
faiss-cpu